% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{CTdata}
\alias{CTdata}
\title{The results of existing clinical trial data organized as a data file, for use in Vignette 3}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 56 rows and 7 columns.
}
\usage{
data(CTdata)
}
\description{
The results of existing clinical trial data organized as a data file, for use in Vignette 3
}
\keyword{datasets}
