% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSimulatedResults.R
\name{generateSimulatedResults}
\alias{generateSimulatedResults}
\title{Generate Simulated Results}
\usage{
generateSimulatedResults(
  trialdesigns,
  respparamsets,
  blparamsets,
  censorparams,
  modelparams,
  simparam,
  analysisparams,
  rawdataout = FALSE,
  lambda_cor = NA,
  n_cores = 1,
  dgp_architecture = "mvn"
)
}
\arguments{
\item{trialdesigns}{A list of trial designs of the form generated
by \link{buildtrialdesigns}}

\item{respparamsets}{A list of options for \code{respparam} as
described in \link{generateData}}

\item{blparamsets}{A list of options for \code{blparam} as
described in \link{generateData}}

\item{censorparams}{A data.table where each row is a set of options for
\code{censorparam} as described in \link{censordata}}

\item{modelparams}{A data.table where each row is a set of options for
\code{modelparams} as described in \link{generateData}}

\item{simparam}{The options that control the logistics of the simulations:
\itemize{
\item{\code{Nreps}}{  Number of repititions for a given set of parameters}
\item{\code{progressiveSave=TRUE}}{  Do you want to save chunks as you go?
Helpful if the run is going to take a long time, and you don't want the
run to be interrupted 8 hours in to a 9 hour run and lose everything.}
\item{\code{basesavename}}{  If you are using progressivesave, what do you
want the base of the filename to be?  Will have _save1, _save2, etc appended}
\item{\code{nRep2save}}{  How many parameter sets to you want to run at
a time before saving?}
\item{\code{saveunit2start}}{  If you got through part of your parameter
space but not all of it, you can start partway through. If starting
from the beginning, set to 1}
\item{\code{savedir}}{  What directory do you want to save to?}
}}

\item{analysisparams}{The options used by \link{lme_analysis}. Has the
following components:
\itemize{
\item{\code{useDE=TRUE}}{  Binary: should the LME model use expectnacy
information if available?}
\item{\code{t_random_slope=FALSE}}{  Binary: should the LME model use
random slopes, as well as random intercepts, for participants?}
}}

\item{lambda_cor}{Correlation decay rate (NA for auto, see
\link{generateData})}

\item{dgp_architecture}{DGP architecture for the biomarker-treatment
interaction (\code{"mvn"} or \code{"mean_moderation"}). See
\link{generateData} for details.}

\item{rawdataout=FALSE}{Binary - output the raw data for all
simulations?  This is memory intensive for large runs.}
}
\value{
Returns a list with three named parts:
\itemize{
\item{\code{results}}{  A large data table that tells you the parameters used
and the analysis results for every simulated trial, one row per simualated trial.}
\item{\code{parameterselections}}{  A list desribing the parameterspace used.
This is important because some of the information in \code{out$results} is
the index of the parameter set selection, and you will need this additional
information to interpret it. For example, \code{trialdesigns}, \code{respparamset},
\code{blparamset}, and \code{modelparamset} are all indexed.}
\item{\code{rawdataout}}{  If you have rawdataout set to \code{TRUE}, you will
also get (1) \code{drawdataout$precensor}, a list of simulated trial data of
a length corresponding to the number of parameter permulations, with the order
parallel to that of the main output results. Repititions of a particular
parameter set are mixed, identified by \code{irep}. And, \code{rawdataout$postcensor},
which has the same format but with the psotcensoring data}
}
}
\description{
\code{generateSimulatedResults} rotates through a parameter space, creating
and analyizing simulated data and collating the results
}
\details{
This is a wrapper for \link{generateData} and \link{lme_analysis} that
systematically works its way through a set of parameters and creates
a specified number of repetitions of simulated data and results of
analyizing that data for each set of parameter specifications.
}
\examples{
  # See vignettes for examples
}
