% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_analysis.R
\name{lme_analysis}
\alias{lme_analysis}
\title{LME Analysis}
\usage{
lme_analysis(trialdesign_set, dat, op)
}
\arguments{
\item{trialdesign_set}{The set of all pathways for the clinical trial}

\item{dat}{A dat file as generated by \link{generateData}}

\item{op}{an options file with the following components:
\itemize{
\item{\code{op$useDE=TRUE}}{  Binary: should the LME model use expectnacy
information if available?}
\item{\code{op$t_random_slope=FALSE}}{  Binary: should the LME model use
random slopes, as well as random intercepts, for participants?}
\item{\code{op$full_model_out=FALSE}}{  Binary: if false, will output model
parameters as a row of numeric values for simulation processing (most common);
if true, will output an array that allows you to examine the results of a single
analysis in more detail}
\item{\code{op$simplecarryover=FALSE}}{  Binary: if true will add "time since
drug" as a free parameter in the model}
\item{\code{op$carryover_t1half=0}}{  If non-zero, will build in an exponential
falloff in the "on drug or not" term in the analysis (Db), using this as the
half-life.}
}}
}
\value{
Generally returns a one row data table with components:
\itemize{
\item{\code{out$beta}}{  The coefficient corresponding to an interaction
between biomarker and outcome}
\item{\code{out$betaSE}}{  Standard error of the above coefficient}
\item{\code{out$p}}{  p-value conveying the statistical significance of the coefficient}
\item{\code{out$issingular}}{  Was the model fit flagged as singular?}
\item{\code{out$warning}}{  Any warnings from \code{fitMsgs}}
}
However, if you have full_model_out set to TRUE, you will instead get an array with components:
\itemize{
\item{\code{form}}{  The model fit by lmer}
\item{\code{fit}}{  The output of lmer}
\item{\code{datamerged}}{  The data in the form it was provided to lmer}
\item{\code{stdout}}{  The standard one row data table that's usually returned when
full_mode_out is set to FALSE, with beta, betaSE, etc}
}
}
\description{
\code{lme_analysis} analyzes the data from a single simulated trial
}
\details{
Analyzes the simulated data (including all paths) from a single simulated
trial. Most often called by \code{generateSimualtedResults}. See vignettes
for additional information.
}
\examples{
# See vignettes 1 & 2 for examples of how to call from the function generateSimualtedResults.R
# See vignette 3 for an example of how to apply lme_analysis to actual clincal trial results
}
