% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{modgompertz}
\alias{modgompertz}
\title{Modified Gompertz}
\usage{
modgompertz(t, maxr, disp, rate)
}
\arguments{
\item{t}{time - no specific units, just make sure it's consistant across your modeling}

\item{maxr}{maximum response that can be attained (ceiling)}

\item{disp}{vertical displacement}

\item{rate}{rate}
}
\value{
A numeric value indicating how much "improvement" there has been since baseline
as of point \code{t} in time. A negative value indicates worsening rather than improvement.
}
\description{
\code{modgompertz} takes response parameters and returns change in symptoms as a function of time
}
\details{
This is a slight modification of a standard gompertz function that is designed to model
how symptoms can change as a function of time.
}
