% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotfactortrajectories}
\alias{plotfactortrajectories}
\title{Plot factor trajectories}
\usage{
plotfactortrajectories(data, tinfo, tds, opt2plot, mergeacrossreps = TRUE)
}
\arguments{
\item{data}{the \code{$rawdata$precensor} component of the output of \link{generateSimulatedResults}
(run with rawdataout=TRUE)}

\item{tinfo}{the \code{$results} component of the output of \link{generateSimulatedResults}}

\item{tds}{the \code{$parameterselections$trialdesigns} component of the output of
\link{generateSimulatedResults}}

\item{opt2plot}{a data.table with a single row using parameter names to instruct the
function which parameters to hold constant and plot the results of}

\item{mergeacrossreps=TRUE}{logical input of whether to merge across repititions}
}
\value{
a ggplot object containing the plot
}
\description{
\code{plotfactortrajectories} plots the averaged output of the factors
used to generate simulated trial data
}
\examples{
\dontrun{
data(results_trajectories)
dat <- results_trajectories$rawdata$precensor
tinfo <- results_trajectories$results
tds <- results_trajectories$parameterselections$trialdesigns
param2hold <- data.table(carryover_t1half = 0, trialdesign = 1)
p <- plotfactortrajectories(dat, tinfo, tds, param2hold)
}
}
