% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagedocumentation.R
\docType{package}
\name{pmsimstats}
\alias{pmsimstats-package}
\alias{pmsimstats}
\title{Personalized medicine stimulation-based statistics}
\description{
The \code{pmsimstats} package is a simple set of tools to
help with implementing simulation-based powercalculations and
bias assessments the ability of different clinical trial
designs to answer personalized medicine (or precision medicine)
oriented hypotheses - specifically, the ability of a
biomarker measured at baseline to predict the biologic
response to a treatment, in the presence of other confounding
effects, including regression to the mean and expectancy
related effects.
}
\section{Further documentation}{

See ___ publication or the two included vignettes, which
generate the analyses and figures from the publication.
}

\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

Authors:
\itemize{
  \item Rebecca C. Hendrickson \email{rebecca.hendrickson@va.gov}
}

}
