% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carryover_analysis.R
\name{print_carryover_summary}
\alias{print_carryover_summary}
\title{Print summary of carryover characterization}
\usage{
print_carryover_summary(result)
}
\arguments{
\item{result}{Output from characterize_carryover()}
}
\description{
Print summary of carryover characterization
}
