% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carryover_analysis.R
\name{print_trial_summary}
\alias{print_trial_summary}
\title{Print summary of extended trial analysis}
\usage{
print_trial_summary(result)
}
\arguments{
\item{result}{Output from analyze_trial_extended()}
}
\description{
Print summary of extended trial analysis
}
