% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reknitsimresults}
\alias{reknitsimresults}
\title{Reknit simulated results}
\usage{
reknitsimresults(basesavename, savedir)
}
\arguments{
\item{basesavename}{The basesavename as described in \link{generateSimulatedResults}}

\item{savedir}{The directory the files are saved in}
}
\value{
The output is a file just like you would have received from
\link{generateSimulatedResults} if it had all been run at once.
}
\description{
\code{reknitsimresults} recombines the results of simulations that have the
exact same parameter space set, but were run as separate chunks - e.g., you
ran 100 repititions, realized you needed more, and ran 150 more to have 250
total. Does not work with rawdataout.
}
