% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{results_maxes}
\alias{results_maxes}
\title{The results of the vigenette Part 1 used to plot maxes respone paramater space}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(results_maxes)
}
\description{
The results of the vigenette Part 1 used to plot maxes respone paramater space
}
\keyword{datasets}
