% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{results_rates}
\alias{results_rates}
\title{The results of the vigenette Part 1 used to plot rate response paramater space}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(results_rates)
}
\description{
The results of the vigenette Part 1 used to plot rate response paramater space
}
\keyword{datasets}
