% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{results_trajectories}
\alias{results_trajectories}
\title{The results of the vigenette Part 1 used to plot the trajectories}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(results_trajectories)
}
\description{
The results of the vigenette Part 1 used to plot the trajectories
}
\keyword{datasets}
